requirements.txt
aniso8601==9.0.1
attrs==21.2.0
autopep8==1.5.7
click==8.0.1
colorama==0.4.4
Flask==1.1.1
Flask-Cors==3.0.10
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
pycodestyle==2.7.0
PyJWT==1.4.2
pyrsistent==0.17.3
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
toml==0.10.2
Werkzeug==0.16.1
